implement extend method for Xpub
