+++ contracts/cw-bitcoin/src/interface.rs
@@ -4,13 +4,13 @@ use bitcoin::BlockHeader;
 use cosmwasm_schema::cw_serde;
 use cosmwasm_schema::schemars::JsonSchema;
 use cosmwasm_schema::serde::{de, ser, Deserialize, Serialize};
-use cosmwasm_std::from_json;
 use cosmwasm_std::to_json_vec;
 use cosmwasm_std::Addr;
 use cosmwasm_std::Binary;
 use cosmwasm_std::Coin;
 use cosmwasm_std::Storage;
 use cosmwasm_std::Uint128;
+use cosmwasm_std::{from_json, StdError};
 use cw_storage_plus::Deque;
 use derive_more::{Deref, DerefMut};
 use sha2::{Digest, Sha256};
@@ -38,7 +38,8 @@ use crate::error::ContractError;
 use crate::error::ContractResult;
 use crate::header::WorkHeader;
 use crate::header::WrappedHeader;
-use crate::utils::add_exp_tweak;
+use libsecp256k1_core::curve::{Affine, ECMultContext, Field, Scalar};
+use libsecp256k1_core::util::{TAG_PUBKEY_EVEN, TAG_PUBKEY_ODD};
 
 #[derive(Deref, DerefMut)]
 pub struct DequeExtension<'a, T> {
@@ -398,12 +399,53 @@ impl Xpub {
         &self.key
     }
 
+    fn parse_pubkey(&self) -> ContractResult<Affine> {
+        let bytes = self.public_key.serialize();
+        let mut x = Field::default();
+        if !x.set_b32(arrayref::array_ref!(&bytes, 1, 32)) {
+            return Err(StdError::generic_err("invalid pubkey").into());
+        }
+        let mut elem = libsecp256k1_core::curve::Affine::default();
+        elem.set_xo_var(&x, bytes[0] == TAG_PUBKEY_ODD);
+        Ok(elem)
+    }
+
+    fn add_exp_tweak(&self, secret: &secp256k1::SecretKey) -> ContractResult<secp256k1::PublicKey> {
+        let tweak = secret.secret_bytes();
+        let mut elem = self.parse_pubkey()?;
+        let mut scala = Scalar::default();
+        if bool::from(scala.set_b32(&tweak)) {
+            return Err(StdError::generic_err("invalid secret").into());
+        }
+
+        let ctx = ECMultContext::new_boxed();
+        let mut r = libsecp256k1_core::curve::Jacobian::default();
+        let a = libsecp256k1_core::curve::Jacobian::from_ge(&elem);
+        let one = libsecp256k1_core::curve::Scalar::from_int(1);
+        ctx.ecmult(&mut r, &a, &one, &scala);
+
+        elem.set_gej(&r);
+
+        let mut ret = [0u8; 33];
+
+        elem.x.normalize_var();
+        elem.y.normalize_var();
+        elem.x.fill_b32(arrayref::array_mut_ref!(ret, 1, 32));
+        ret[0] = if elem.y.is_odd() {
+            TAG_PUBKEY_ODD
+        } else {
+            TAG_PUBKEY_EVEN
+        };
+        let pubkey = secp256k1::PublicKey::from_slice(&ret)?;
+        Ok(pubkey)
+    }
+
     /// Deterministically derive the public key for a signatory in a signatory set,
     /// based on the current signatory set index.
     pub fn derive_pubkey(&self, sigset_index: u32) -> ContractResult<secp256k1::PublicKey> {
         let child_number = bitcoin::util::bip32::ChildNumber::from_normal_idx(sigset_index)?;
         let (sk, _) = self.ckd_pub_tweak(child_number)?;
-        add_exp_tweak(&self.public_key, &sk)
+        self.add_exp_tweak(&sk)
     }
 }
